Interface to create a VQ (or, an RX context)
Now, we just need a way to associate a VQ with a TX context.  A TX
context has a single VQ, and a VQ can be shared by many TX contexts.
If we implement skb->mark classification, then all these become much
simpler, but we chose to do it this way for now.