VQ bug: ignore non-active VQs when reassigning capacities
This is a serious bug: since non-backlogged VQs were assigned their full
rate, it means that a burst of packets arriving would be drained at
full VQ rate before being assigned a new capacity.  This would lead to
a temporal state where net RX rate is higher than VQ drain rate, which
is not good.  This commit mitigates the issue to a certain extent, but
not completely.