Hacky way to make rate control's counters per-cpu.
per_cpu_alloc allocates in GFP_KERNEL context that can sleep.  But
we need this allocation in softirq context, where we cannot sleep.
Hence this hackery.

Is this needed?

We need per-cpu counters as these get modified for every packet.
If we operate under Xen, it's likely that ALL packets received by
a VM land on one CPU, or at least, all packets to the same destination
land on one CPU, so this could be over overly optimising it.

Also, if the traffic is largely one-way, the reverse traffic is
small and infrequent enough that the per-cpu counters might not
matter.

But it's best to have it.  The only downside is that we're wasting
some memory.