Remove memory allocations from datapath
This means more complicated memory management, but it keeps the
datapath free from unnecessary GFP_ATOMIC memory allocations.  An
alternative design is to queue an allocation work item for a worker
thread to do it, but it can be more complicated, as we have to still
"allocate" a work item.