bug: fix token refresh check condition
First, this condition used to be a time-based check.  Later,
I changed it to a token based check, which could actually
deadlock!

A deadlock happens when the per-cpu queue's head packet size
is greater than the local tokens it has, but the total tokens
the queue has is less than the minimum burst that a per-cpu
queue borrows.

This can be fixed either by borrowing as much tokens as
available, or by checking the condition keeping borrowing
in mind.